+++ src/Application/NoActionMidiFileSaver.cs
@@ -1,9 +1,0 @@
-namespace MidiRecorder.Application;
-
-public class NoActionMidiFileSaver<TMidiEvent> : IMidiFileSaver<TMidiEvent>
-{
-    public void Save(IEnumerable<TMidiEvent> events, string filePath, int timeDivision)
-    {
-        Console.WriteLine("FAKE SAVE: " + filePath);
-    }
-}